% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correct.conc.R
\name{correct.conc}
\alias{correct.conc}
\title{correct.conc}
\usage{
correct.conc(sample.database = NULL, std.coefs = NULL,
  blank.vite = NULL, extraction.vol.ml = NULL,
  filtration.vol.l = NULL)
}
\arguments{
\item{sample.database}{character string name or filepath and name of DATABASE name (or path to database if not in working directory) output from id.peaks function}

\item{std.coefs}{character string name or filepath and name of a Standard Curve RSQLite database}

\item{blank.vite}{value of the AVERAGE blank vitamin e peak area (need to integrate peaks for your blank samples and use view.database function to manually get and make average)}

\item{extraction.vol.ml}{volume of the extraction solvent used for pigment extractions - in MILLILITRES}

\item{filtration.vol.l}{volume of the amout of sample filtered for extraction - in LITRES}
}
\value{
'correct.conc' table within main sample database with concentrations for pigments a standard is available for in ug per litre, or the ratio of pigment area to chlorophyll area for those pigments a standard is not available for.
}
\description{
calculate pigment concetrations using std coeffients
}
