% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify.peaks.R
\name{identify.peaks}
\alias{identify.peaks}
\title{identify.peaks}
\usage{
\method{identify}{peaks}(my.data = "", nups.451 = 10, nups.223 = 20)
}
\arguments{
\item{my.data}{character string name or filepath and name of the CSV containing the original data}

\item{nups.451}{number of ups to define peak at 451 nm (can be altered if peaks not captured correctly)}

\item{nups.223}{number of ups to define peak at 223 nm (can be altered if peaks not captured correctly)}
}
\value{
SQLite database in current working directory updated to output of integrate.peaks() with designated peak identifications determined during interactive peak ID shinyApp.
}
\description{
identify peaks in output of EXPORT3D.mac for ChemStation hplc data
}
