% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/integrate.peaks.R
\name{integrate.peaks}
\alias{integrate.peaks}
\title{integarte.peaks}
\usage{
integrate.peaks(my.data = "", blank = NULL, nups.451 = 10,
  nups.223 = 20)
}
\arguments{
\item{my.data}{character string name or filepath and name of the CSV containing the data}

\item{blank}{character string name or filepath and name of a single representative blank sample CSV}

\item{nups.451}{number of ups to define peak at 451 nm (can be altered if peaks not captured correctly)}

\item{nups.223}{number of ups to define peak at 223 nm (can be altered if peaks not captured correctly)}
}
\value{
SQLite database containing the following tables:
   1. meta.data: meta data table containing sample information, including information on extraction and filtration volumes and cell counts associated with the samples.
   2. vite_blank_data: containing associated blank samples vitamin E concentrations
   3. mobile_phase_reference_spectra: full spectra for blank samples used in blanking process
   4. library_meta_data: meta data table for reference library showing pigment names, wavelength observed, retention times
   5. library_abs_data: full spectra for all pigments in the library
   6. library_vite_abs_data: spectra for vitamin E internal standard library
   7. raw_abs_data: raw EXPORT3D.mac output from ChemStation
   8. corrected_abs_data: 223 nm, 431 nm, and 451 nm absorbance data, corrected for background (mobile phase reference) and 751 nm background absorbance
   9.etc
}
\description{
identify peaks in output of EXPORT3D.mac for ChemStation hplc data
}
