% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.database.R
\name{view.database}
\alias{view.database}
\title{view.database}
\usage{
view.database(my.data = "")
}
\arguments{
\item{database}{character string name or filepath and name of database}
}
\description{
view database contents made by ProcessHplc package functions
}
